requirements.txt
bcrypt==4.0.1
black==22.10.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
humanize==4.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
pathspec==0.10.2
platformdirs==2.5.4
protobuf==3.20.1
PyYAML==6.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
